Define separate test services in reactive OpenShift scenario
The reactive OpenShift scenario was failing otherwise.
