+++ nosql-db/mongodb-reactive/src/test/java/io/quarkus/ts/nosqldb/mongodb/OpenShiftMongoDbIT.java
@@ -1,7 +1,23 @@
 package io.quarkus.ts.nosqldb.mongodb;
 
+import io.quarkus.test.bootstrap.MongoDbService;
+import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.services.Container;
+import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
-public class OpenShiftMongoDbIT extends MongoDbIT {
+public class OpenShiftMongoDbIT extends AbstractMongoDbIT {
+
+    @Container(image = "${mongodb.image}", port = 27017, expectedLog = "Waiting for connections")
+    static MongoDbService database = new MongoDbService();
+
+    @QuarkusApplication
+    static RestService app = new RestService()
+            .withProperty("quarkus.mongodb.connection-string", database::getJdbcUrl);
+
+    @Override
+    protected RestService getApp() {
+        return app;
+    }
 }
